// src/js/features/actions/types.ts
export enum Actions {
  SET_ONLINE_MODE = 'SET_ONLINE_MODE',
  TOGGLE_DISPLAY_OPTIONS = 'TOGGLE_DISPLAY_OPTIONS',
  TOGGLE_FONT_OPTIONS = 'TOGGLE_FONT_OPTIONS',
  TOGGLE_DARK_MODE = 'TOGGLE_DARK_MODE',
  TOGGLE_TRANSLATION_OPTIONS = 'TOGGLE_TRANSLATION_OPTIONS',
  TOGGLE_TRANSLITERATION_OPTIONS = 'TOGGLE_TRANSLITERATION_OPTIONS',
  TOGGLE_LARIVAAR_OPTION = 'TOGGLE_LARIVAAR_OPTION',
  TOGGLE_UNICODE_OPTION = 'TOGGLE_UNICODE_OPTION',
  TOGGLE_SPLIT_VIEW_OPTION = 'TOGGLE_SPLIT_VIEW_OPTION',
  TOGGLE_PARAGRAPH_VIEW_OPTION = 'TOGGLE_PARAGRAPH_VIEW_OPTION',
  SET_UNICODE = 'SET_UNICODE',
  SET_TRANSLATION_LANGUAGES = 'SET_TRANSLATION_LANGUAGES',
  SET_DARK_MODE = 'SET_DARK_MODE',
  SET_TRANSLITERATION_LANGUAGES = 'SET_TRANSLITERATION_LANGUAGES',
  SET_FONT_SIZE = 'SET_FONT_SIZE',
  TOGGLE_LARIVAAR_ASSIST_OPTION = 'TOGGLE_LARIVAAR_ASSIST_OPTION',
}